Remove dead code, match other node styling
