+++ packages/workflow-diagram/src/nodes/EmptyWorkflowNode.tsx
@@ -1,0 +1,7 @@
+import React, { memo } from "react";
+
+const EmptyWorkflowNode = () => <div></div>;
+
+EmptyWorkflowNode.displayName = "WorkflowNode";
+
+export default memo(EmptyWorkflowNode);
